Added more logging/output.

+++ TerrariaPatcher/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
@@ -13,26 +14,56 @@ namespace TerrariaPatcher
         static readonly string TmpDir = "apitmp";
         static void Main(string[] args)
         {
-            Environment.CurrentDirectory = @"C:\Program Files (x86)\Steam\steamapps\common\terraria";
-            if (File.Exists("Terraria.exe"))
+            try
             {
+                if (!File.Exists("Terraria.exe"))
+                {
+                    Output("Terraria.exe not found");
+                    Console.ReadLine();
+                    return;
+                }
+
                 string md5;
                 using (var fs = new FileStream("Terraria.exe", FileMode.Open, FileAccess.Read, FileShare.Read))
                 {
                     md5 = MD5(fs);
                 }
+                var terrariaver = FileVersionInfo.GetVersionInfo("Terraria.exe");
+                string apitmp = Path.Combine(Environment.CurrentDirectory, "apitmp");
+
+                Output("Downloading Patches Information");
+                var patches = DownloadPatches();
+
+                string vmd5;
+                if (!patches.TryGetValue(terrariaver.ProductVersion, out vmd5))
+                {
+                    Output("Terraria version not supported ({0})", terrariaver.ProductVersion);
+                    Console.ReadLine();
+                    return;
+                }
+
+                if (vmd5 != md5)
+                {
+                    Output("Terraria hash mismatch (already patched?).");
+                    Output("Version {0}", terrariaver.ProductVersion);
+                    Output("Expected {0} got {1}", vmd5, md5);
+                    Console.ReadLine();
+                    return;
+                }
 
                 if (!Directory.Exists(TmpDir))
                     Directory.CreateDirectory(TmpDir);
 
+                Output("Downloading Diff");
                 string patch = DownloadPatch(md5);
                 File.WriteAllText(Path.Combine(TmpDir, "Terraria.diff"), patch);
 
+                Output("Extracting Resources");
                 File.WriteAllBytes(Path.Combine(TmpDir, "ildasm.exe"), Properties.Resources.ildasm);
                 File.WriteAllBytes(Path.Combine(TmpDir, "IlDasmrc.dll"), Properties.Resources.IlDasmrc);
                 File.WriteAllBytes(Path.Combine(TmpDir, "patch.exe"), Properties.Resources.patch);
 
-
+                Output("Running IlDasm");
                 var proc = new Process();
                 proc.StartInfo.FileName = "apitmp/ildasm.exe";
                 proc.StartInfo.Arguments = "Terraria.exe /output:" + Path.Combine(TmpDir, "Terraria.il");
@@ -44,10 +75,11 @@ namespace TerrariaPatcher
                 Log(proc.StandardOutput.ReadToEnd());
                 Log(proc.StandardError.ReadToEnd());
 
+                Output("Running Patch");
                 proc = new Process();
                 proc.StartInfo.FileName = "apitmp/patch.exe";
                 proc.StartInfo.Arguments = "-u -o NewTerraria.il -i Terraria.diff";
-                proc.StartInfo.WorkingDirectory = Path.Combine(Environment.CurrentDirectory, "apitmp");
+                proc.StartInfo.WorkingDirectory = apitmp;
                 proc.StartInfo.UseShellExecute = false;
                 proc.StartInfo.RedirectStandardOutput = true;
                 proc.StartInfo.RedirectStandardError = true;
@@ -56,12 +88,14 @@ namespace TerrariaPatcher
                 Log(proc.StandardOutput.ReadToEnd());
                 Log(proc.StandardError.ReadToEnd());
 
+                if (File.Exists(Path.Combine(apitmp, "NewTerraria.il.rej")))
+                    Log("\n\n[[{0}]]\n\n",File.ReadAllText(Path.Combine(apitmp, "NewTerraria.il.rej")));
 
-
+                Output("Running Ilasm");
                 proc = new Process();
                 proc.StartInfo.FileName = Path.Combine(RuntimeEnvironment.GetRuntimeDirectory(), "ilasm.exe");
                 proc.StartInfo.Arguments = "NewTerraria.il /quiet /output=Terraria.exe /res=Terraria.res";
-                proc.StartInfo.WorkingDirectory = Path.Combine(Environment.CurrentDirectory, "apitmp");
+                proc.StartInfo.WorkingDirectory = apitmp;
                 proc.StartInfo.UseShellExecute = false;
                 proc.StartInfo.RedirectStandardOutput = true;
                 proc.StartInfo.RedirectStandardError = true;
@@ -70,36 +104,74 @@ namespace TerrariaPatcher
                 Log(proc.StandardOutput.ReadToEnd());
                 Log(proc.StandardError.ReadToEnd());
 
+                Output("Finishing");
                 if (File.Exists(Path.Combine(TmpDir, "Terraria.exe")))
                 {
                     File.Copy(Path.Combine(TmpDir, "Terraria.exe"), "Terraria.exe", true);
-                    Console.WriteLine("Success");
+                    Output("Success");
                 }
                 else
                 {
-                    Console.WriteLine("Failed");
+                    Output("Failed");
                 }
-
-                Directory.Delete(TmpDir, true);
             }
-            else
+            catch (Exception ex)
             {
-                Console.WriteLine("Terraria.exe not found");
+                Log(ex.ToString());
+                Output("Exception: " + ex.ToString());
             }
+            finally
+            {
+                if (Directory.Exists(TmpDir))
+                    Directory.Delete(TmpDir, true);
+            }
+
             Console.ReadLine();
+        }
+
+        static void Output(string str)
+        {
+            if (str.Length < 1)
+                return;
+            Console.WriteLine(str);
+            Log(str);
+        }
+        static void Output(string str, params object[] objs)
+        {
+            Output(string.Format(str, objs));
         }
 
         static void Log(string str)
         {
             if (str.Length < 1)
                 return;
-            File.AppendAllText("patchlog.txt", str);
+            File.AppendAllText("patchlog.txt", str + "\n");
+        }
+        static void Log(string str, params object[] objs)
+        {
+            Log(string.Format(str, objs));
+        }
+
+        static Dictionary<string, string> DownloadPatches()
+        {
+            var ret = new Dictionary<string, string>();
+
+            string patches = new WebClient().DownloadString("http://dl.dropbox.com/u/29760911/TerrariaApi/patches.txt");
+            var ps = patches.Split('\n', '\r');
+            foreach (var p in ps)
+            {
+                var kv = p.Split('|');
+                if (kv.Length != 2)
+                    continue;
+
+                ret.Add(kv[0], kv[1]);
+            }
+            return ret;
         }
 
         static string DownloadPatch(string md5)
         {
-            return File.ReadAllText(@"C:\Program Files (x86)\Steam\steamapps\common\terraria\TerrariaAPI\Terraria\test\Terraira_88C4FEFB311D0563CBA46D1720708ADA.diff");
-            //return new WebClient().DownloadString("http://shankshock.com/Terraria_" + md5 + ".diff"); ;
+            return new WebClient().DownloadString("http://dl.dropbox.com/u/29760911/TerrariaApi/Terraira_" + md5 + ".diff");
         }
 
         static string MD5(Stream stream)
